Rename SantanderConsumerBankAGPDFExtractor to SantanderConsumerBankPDFExtractor
